Android: SVG Images not rendering the first time a view is displayed
Since SVG images are loaded asynchronously, we need to invalidate the
view after drawing to the canvas so that the view will be re-rendered.

Storing a reference to the actual SvgView in ViewManager, and passing a
reference to the ViewManager to each shadow node, so that shadow nodes
can invalidate the root SVG view.